Fixed ToZapLogPlugin not generic enough
The to zap log plugin was not generic enough turning only "EOSIO" log level to their correct
level. The component has been generalized so it can be configured to support any binary
format.

The actual implementation of the superviser will decide how to configure the plugin to
make it work like it's supposed to do.