Added commands logging through info and move deep mind debug to info
Debugging deep mind if often done when there is a problem on production node. Since it
was in debug, the command used to enabled deep mind logging seemed to not work but it was
not writing anything since it was logging in debug which is not active on production node.

By using info instead, the command will work right away without having to tweak the logging
level first. Debugging deep mind is usually used with a command that enables it for a
few seconds only, so we should be good.