Local files ready to upload are now stored uncompressed and compression is done when remote uploading
The local file write is always being performed sequentially one file at a time acting a bit like a write ahead log. On the counter part, the remote upload is done with mutliple uploaders in parallel meaning we have better throughput and less contention there.

By doing the compression only on the remote upload, we get a 50% gain on amout of `blocks/s` we are able to do which is quite interesting. This is at the expense of much bigger blocks being written to disk since now uncompressed. Shouldn't be that bad but could be consider or we could also put a limit on how many bytes can live on the local disk before starting to block.

We could also think about looking for a real write ahead log library like https://github.com/tidwall/wal and at the same design re-design the reader-node to ensure we are blocking until the write ahead log is correctly written. Requires heavier refactoring that is not super important right now.